refactor - move all scripts manager files into its own sub folder decalre file paths as consts in file_utils
