+++ src/commands_automator/scripts_manager/config/commands-executor-config.json
