+++ src/commands_automator/scripts_manager/config/scripts_config.json
