+++ src/commands_automator/scripts_manager/user_scripts/cmd/fast_copy.cmd
