+++ src/commands_automator/scripts_manager/user_scripts/cmd/kill_process_holding_file.cmd
