+++ src/commands_automator/scripts_manager/user_scripts/cmd/show_free_space_drive_c.cmd
