+++ src/commands_automator/scripts_manager/user_scripts/python/remove_parenthesis.py
