+++ src/commands_automator/scripts_manager/user_scripts/python/replace_chars.py
