+++ src/commands_automator/scripts_manager/user_scripts/shell/convert.sh
