+++ src/commands_automator/scripts_manager/user_scripts/shell/double_the_slash.sh
